requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
openai
sentence-transformers
faiss-cpu
numpy